requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.2
cssselect==1.1.0
Django==3.1.1
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-js-asset==1.2.2
djangorestframework==3.11.1
isort==5.5.3
lazy-object-proxy==1.4.3
lxml==4.5.2
Markdown==3.2.2
markdown2==2.3.9
mccabe==0.6.1
Pillow==7.2.0
pylint==2.6.0
pyquery==1.4.1
pytz==2020.1
readtime==1.1.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
